Render the review summary inside Layout's content cell
Replace the body-level fixed portal with a persistent overlay slot on Layout.
The review layer now renders in slotOverlay, an always-mounted cell that
occupies the grid's content area (the full container on mobile), so the CSS
grid supplies the sidebar/panel insets that ReviewSummaryPortal previously
re-derived by hand. This deletes useDesktopChromeInsets, the duplicated
600px breakpoint matchMedia listener, the -10000px off-screen parking hack,
and the legacy portal-host cleanup. The hidden state is reduced to
visibility: hidden plus inert, which was verified empirically to keep
thumbnail iframe documents alive without reloading.
