+++ code/core/src/manager/components/review/components/ReviewPersistentLayer.tsx
@@ -1,6 +1,6 @@
 import React, { type FC } from 'react';
 
-import { ReviewSummaryPortal } from '../screens/ReviewSummaryPortal.tsx';
+import { ReviewSummaryHost } from '../screens/ReviewSummaryHost.tsx';
 import { useReviewNavigationInterceptor } from '../useReviewNavigationInterceptor.ts';
 import { useReviewShortcuts } from '../useReviewShortcuts.ts';
 import { ReviewNotification } from './ReviewNotification.tsx';
@@ -13,14 +13,14 @@ const ReviewNavigationLayer: FC = () => {
   return (
     <>
       <ReviewNotification />
-      <ReviewSummaryPortal />
+      <ReviewSummaryHost />
     </>
   );
 };
 
 /**
- * Always-mounted review layer. Hosts the review state provider, navigation
- * interceptor/shortcuts, and the summary portal so review survives story
+ * Always-mounted review layer, rendered in the Layout's overlay slot. Hosts the review state
+ * provider, navigation interceptor/shortcuts, and the summary host so review survives story
  * navigation. Self-gates: the provider stays dormant until a review is pushed.
  */
 export const ReviewPersistentLayer: FC = () => (
